requirements.txt
streamlit
langchain_community 
pypdf 
faiss-cpu
sentence-transformers
tf-keras
faiss-cpu